pyproject.toml
[project]
name = "codingengine"
version = "0.1.0"
description = "编排+执行+迭代的 AI 辅助开发工具，不污染项目目录"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-bdd", "black", "ruff"]

[project.scripts]
codingengine = "codingengine.cli:main"
ce = "codingengine.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
